[Miguel_Benjamín_Hernan_Matias]-[Fakers aplicados en seeder]
Se aplican mediante los seeders los fakers

+++ proyectoDBD/database/migrations/2014_10_12_000000_create_users_table.php
@@ -18,7 +18,7 @@ class CreateUsersTable extends Migration
             $table->string('nombre');
             $table->string('apellido');
             $table->string('contraseña');
-            $table->integer('numeroTelefono');
+            $table->BigInteger('numeroTelefono');
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->rememberToken();
